requirements.txt
streamlit>=1.32.0
google-generativeai>=0.3.2
protobuf>=3.20.3
emoji>=2.2.0